requirements.txt
fastapi==0.100.0           
uvicorn==0.23.1            
SQLAlchemy==2.0.35          
pydantic==2.10.2          
python-jose==3.3.0         
passlib[bcrypt]==1.7.4     
psycopg2-binary==2.9.7     
python-dotenv==1.0.0       
websockets==11.0.3         
alembic==1.10.2  
email-validator==2.2.0
requests==2.31.0
PyJWT==2.8.0
pyTelegramBotAPI==4.20.0
fastapi-mail==1.4.2
redis==5.0.1
celery==5.3.6
python-multipart==0.0.17
